Address-87: Remove bypass of signature verification (#111)
There is a strong chance that an implementer could get this wrong. This is not required for interoperability and, therefore, is an implementation detail. While the guidance is likely useful in a small number of use cases, the risk to most use cases of this approach being implemented badly and introducing a vulnerability outweighs the value in the applications where it is helpful. As a result it seems best to remove this detail. This may create an un-satisfied security requirement, but as above this appears to be an implementation detail.